moved cache routes and added main content header
