+++ app/routes/admin+/cache/lru.$cacheKey.ts
