+++ app/routes/admin+/cache/route.tsx
