+++ app/routes/admin+/cache/sqlite.$cacheKey.ts
