+++ app/routes/admin+/index.tsx
@@ -1,0 +1,9 @@
+import { ContentHeader, MainContentContainer } from '#app/components/layout'
+
+export default function LayersIndexRoute() {
+	return (
+		<MainContentContainer>
+			<ContentHeader>Welcome Admin</ContentHeader>
+		</MainContentContainer>
+	)
+}
